badminton-analyzer/requirements.txt
anthropic>=0.40.0
flask>=3.0.0
opencv-python-headless>=4.9.0
python-dotenv>=1.0.0
